Kali tools to install + upgrades
When using the free Play version of Offensive Security Proving Grounds, I've found that the instance of kali in the Kali Browser is reset each time. I run this script to have a base version of Kali with the tools I like to use. I also use this if I need to reinstall Kali and want to update, upgrade, and install the tools I like to use.